Unclude complete and working
The unclude.py script is finished.

Unclude is run from the root directory and will find all .c and .h files
that have includes that are not being used by the program. For this to
work there has to be some extra syntax on the include lines.

#incluce <something>//A comma delimited list of what is being used

If it finds an unused include it will notify the user of the file, and
line of that bad include.