Uses Default Rule for Function Call
The default reduction on object when Bison sees the '(' in
a function call is what we want anyway, so I surpressed the output
of conflicts on that case to temporarily solve the problem.

If this is the only conflict we ignore from now on, then we
will have no problems building the tree.