Eliminates S/R Conflict For Good
I added line number tracking to the parser and solved
the S/R conflict by defining a special type of action
for objects that never call a function. This was actually
needed for function definitions because the default
reduce behavior in the S/R actually conflicted with function
calling.

Also, slices work as the range token wasnt being consumed
due to flex thinking 1..5 was two real numbers 1.0 and 0.5
The downside is that we must make all real numbers use digits
on either side of the decimal, or use a different range token.