Adds HklHash and HklTree
Also, I added a new macro for freeing objects
allocated with hkl_alloc_object: hkl_free_object.

Its not mandatory to use. But it could potentially
save trouble to use it when you can.

The current hkl_test.c contains some test code for the
hash table. It seems verbose, but the string input
for the program enters the parser already in HklString
form.

By default HklHash and HklTree return HklPair objects
as their result. This is essentially the same as how
a stl map functions.

HklHash also does not automatically exapand yet.
This will be added later, but won't change the API.