Deque
Deque is now working as an array, push back, pop front work as expected.

Internally the array is resized and setup so that after a resize the
whole array is reordered with the front at index 0 and the back at the
last position. This means that after a resize all insertions will work
as intended.
The grow ability only happens as a last resort. It is only triggered in
an overflow situation.
One issue it does not handle is if it was unable to grow.