Deque Finished
Deque is finished for now.

It uses an array as the internal structure. This gives the following
advantages.
We can mitigate the need to allocate after every push, and deallocate
after every pop.
We gain the reusability of nodes this way

We can initialize the queue to any arbitrary size using the
hkl_deque_new_sized.
This means that if given a relatively ballparked estimate for the memory
needs of a queue
we are able to minimize the calls to malloc to 1 or slightly more.

We are able to use the deque to implement internal arrays in our
langague. They are dynamically sized and therefore will allow HKL to use
them as much as needed.

We can locate the nth item in the array given an index such that -size
<= index < size.

We can locate the nth item in the array given that the queue isn't
empty. This means that cyclical arrays are now as easy to implement as a
single for loop with no extra work required.